requirements.txt
Jinja2>=3.1.4
PyYAML>=6.0.2
invoke>=2.2.0
